refactor(lib): move all utils files into lib folder
I previously had this split up. This shouldn't be the case
