+++ src/pages/api/netflix/added-titles.ts
@@ -1,7 +1,7 @@
 import { type NextApiRequest, type NextApiResponse } from "next";
 
-import { fetchNewTitles } from "@/utils/netflix";
-import { convertDateQueryParam, getFormattedDate } from "@/utils/date";
+import { fetchNewTitles } from "@/lib/netflix";
+import { convertDateQueryParam, getFormattedDate } from "@/lib/date";
 import authenticateRequest from "@/server/authenticateRequest";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
