+++ src/pages/api/post-to-twitter.ts
@@ -2,7 +2,7 @@ import { type NextApiRequest, type NextApiResponse } from "next";
 import { verifySignature } from "@upstash/qstash/nextjs";
 
 import authenticateRequest from "@/server/authenticateRequest";
-import { fetchDailyTitles } from "@/utils/netflix";
+import { fetchDailyTitles } from "@/lib/netflix";
 import { buildThread, createTweetThread } from "@/server/twitter";
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
